☀️ 73% – adopt the battle command action executor
main:0808e680 (748 bytes, BattleCommand_ExecuteSelectedAction) compiles
byte-exact once the battle work fields are reached through the shared
struct instead of absolute addresses, the action definition is fetched a
second time for the cost, and the 0x920 message id is a symbol so it stays
a pool word.

The function lives in translation unit unit-0808e23c, whose members must
agree on the event lookup's prototype. The executor needs the group
parameter declared s32 (a u16 parameter adds a truncation), so the unit's
declarations move to s32 and the one exact caller that relied on the
truncation spells it as a cast; both stay byte-exact.

+++ games/gs1/src/battle/effects/has_matching_event5.c
@@ -1,9 +1,9 @@
 #include "effect_runtime.h"
 
-s32 Func_0808e4b4(s32, u16, void *);
+s32 Func_0808e4b4(s32, s32, void *);
 
 s32 BattleEffect_HasMatchingEvent5(s32 effectId) {
     s32 local;
-    s32 result = Func_0808e4b4(0x70000005, effectId, &local);
+    s32 result = Func_0808e4b4(0x70000005, (u16)effectId, &local);
     return (u32)((-result) | result) >> 31;
 }
